Size is not grandeur

+++ 0x00-hello_world/6-size.c
@@ -5,13 +5,13 @@
  * main - This is a program that print a line with print function
  * Return:0.)
  */
+
 int main(void)
 {
-
-prtintf("Size of Char: %zu byte(s)\n", (unsigned long)sizeof(char));
-prtintf("Size of int: %zu byte(s)\n", (unsigned long)sizeof(int));
-prtintf("Size of long int: %zu byte(s)\n", (unsigned long)sizeof(long int));
-prtintf("Size of long long int: %zu byte(s)\n", (unsigned long)sizeof(long long int));
-prtintf("Size of Float: %zu byte(s)\n", (unsigned long)sizeof(float));
+prtintf("Size of Char: %zu byte(s)\n", sizeof(char));
+prtintf("Size of int: %zu byte(s)\n", sizeof(int));
+prtintf("Size of long int: %zu byte(s)\n", sizeof(long int));
+prtintf("Size of long long int: %zu byte(s)\n", sizeof(long long int));
+prtintf("Size of Float: %zu byte(s)\n", sizeof(float));
 return (0);
 }
